refactor: rename Helpers dir into typings dir and its namespace"
