+++ src/interfaces/GdriveInterface.php
@@ -2,8 +2,8 @@
 
 namespace Yaza\LaravelGoogleDriveStorage\interfaces;
 
-use Yaza\LaravelGoogleDriveStorage\Helpers\GdriveFile;
-use Yaza\LaravelGoogleDriveStorage\Helpers\GdriveFileInfo;
+use Yaza\LaravelGoogleDriveStorage\typings\GdriveFile;
+use Yaza\LaravelGoogleDriveStorage\typings\GdriveFileInfo;
 use \League\Flysystem\DirectoryListing;
 use \League\Flysystem\StorageAttributes;
